% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolagetratings-package.R
\docType{package}
\name{bolagetratings-package}
\alias{bolagetratings}
\alias{bolagetratings-package}
\title{bolagetratings: Ratings for Systembolaget}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Scrapes, merges, and displays Vivino and Untappd ratings for products in Systembolaget.
}
\author{
\strong{Maintainer}: Alexander Thorén \email{alethoren@gmail.com}

}
\keyword{internal}
