% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systembolaget.R
\name{get_products}
\alias{get_products}
\title{Gets all Systembolaget products}
\usage{
get_products(storeId = NULL)
}
\arguments{
\item{storeId}{an optional store id}
}
\value{
data.frame
}
\description{
If a `storeId` is provided, only that stores products are returned
}
