% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systembolaget.R
\name{get_produts_page}
\alias{get_produts_page}
\title{Gets a single page of Systembolaget products}
\usage{
get_produts_page(page = 1, storeId = NULL)
}
\arguments{
\item{page}{the page to get}

\item{storeId}{an optional store id}
}
\value{
data.frame
}
\description{
If a `storeId` is provided, only `productId` and `storeId` are included in
the result.
}
