% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systembolaget.R
\name{get_stores}
\alias{get_stores}
\title{Gets Systembolagets stores and relevant information}
\usage{
get_stores()
}
\value{
data.frame
}
\description{
Gets Systembolagets stores and relevant information
}
