% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge Systembolaget with Vivino ratings}
\usage{
merge_data(systembolaget, vivino)
}
\arguments{
\item{systembolaget}{Systembolaget data, result of [get_products()]}

\item{vivino}{Vivino data, result of [cache_systembolaget_wines()] and
[extract_vivino_products()]}
}
\value{
merged data.frame
}
\description{
Merge Systembolaget with Vivino ratings
}
